CGFloat and Double interchangable

+++ BallroomBuySell/Screens/Shared/SaleItem/Cells/ImageTableCell/ImageTableCell.swift
@@ -23,7 +23,7 @@ class ImageTableCell: UITableViewCell, TableCellProtocol, UICollectionViewDataSo
     @IBOutlet weak var collectionViewHeight: NSLayoutConstraint!
     
     private var maxImageCount = 0
-    private let imageWidth = 800
+    private let imageWidth = 800.0
     private var imagesList = [Data]()
     private var isEditable = true
     var delegate: (ImageCellDelegate & UIViewController)?
@@ -92,7 +92,7 @@ class ImageTableCell: UITableViewCell, TableCellProtocol, UICollectionViewDataSo
         }
         
         let normalizedImage = selectedImage.normalizedImage()
-        let resizedImage = normalizedImage.resize(newWidth: CGFloat(imageWidth))
+        let resizedImage = normalizedImage.resize(newWidth: imageWidth)
         
         picker.dismiss(animated: true, completion: nil)
         if let imageData = resizedImage.pngData() {
